#381 create transformers.py move functionality over to it
